aws-gpt-api/requirements.txt
openai==1.63.2
flask==3.0.3
python-dotenv==1.0.1